+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Versioning].
 ## [Unreleased]
 
 - Fix `required` value for "configuration-path" input. ([#100])
+- Fix interpretation of "dry-run" input in config file. ([#103])
 
 ## [0.2.1] - 2020-03-02
 
@@ -54,3 +55,4 @@ Versioning].
 [#92]: https://github.com/ericcornelissen/svgo-action/pull/92
 [#97]: https://github.com/ericcornelissen/svgo-action/pull/97
 [#100]: https://github.com/ericcornelissen/svgo-action/pull/100
+[#103]: https://github.com/ericcornelissen/svgo-action/pull/103